Prevent crash when trying to load an unsupported language
Someone reported an issue regarding this on the Discord server a few days
ago, and I decided that I would investigate.

On a side note, the CoffeeTranslation codebase is easily the worst part of
the Rehike codebase. It's almost bad enough to warrant being completely
rewritten (again), but I'm not sure if I want to do that.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>