Fix collaborators on browse video renderers
This fixes the other half of #686, resolving the "issue" part of that
issue. Further work will be conducted in the future to make it possible to
display all collaborators of a video on the watch page, but I am currently
tired and this is rushed work which needs to be cleaned up anyway.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>