Safe Twig i18n HTML interpolation implementation (#689)
This replaces the old strategy of trusting the entire localized string when HTML interpolation was required. Only arguments passed with the |safeHtml filter (converted to SafeHtml objects) are trusted.

There was already an instance that I encountered of formatting being missed, albeit benign. Single quote characters should be escaped with the HTML entity "&#039;", but they weren't in the old case.

Because this is not compatible with format(), I added a new method to the i18n delegate. I don't know how to get the current Twig autoescaping strategy, and the function only escapes untrusted content as HTML. It could possibly be done with a Twig extension, but I think that's a bit overkill.